Moved vue 2 sample app to own folder.
