+++ .gitignore
@@ -1,6 +1,6 @@
 .DS_Store
 node_modules
-dist
+demo*
 
 
 # local env files
