--- vue2/src/idsrvAuth-async.d.ts
@@ -1,1 +1,0 @@
-export declare function configureOidc(): Promise<import("./lib/oidc-auth").OidcAuth>;
