--- vue2/src/idsrvAuth.d.ts
@@ -1,2 +1,0 @@
-declare const idsrvAuth: import("./lib/oidc-auth").OidcAuth;
-export default idsrvAuth;
