+++ vue2/src/oidc-auth.d.ts
