+++ vue2/src/sample-app/App.vue
