+++ vue2/src/sample-app/idsrvAuth-async.d.ts
@@ -1,0 +1,1 @@
+export declare function configureOidc(): Promise<import("../oidc-auth").OidcAuth>;
