+++ vue2/src/sample-app/idsrvAuth.d.ts
@@ -1,0 +1,2 @@
+declare const idsrvAuth: import("../oidc-auth").OidcAuth;
+export default idsrvAuth;
