+++ vue2/src/sample-app/main-async.d.ts
