+++ vue2/src/sample-app/main-async.ts
