+++ vue2/src/sample-app/main.d.ts
