+++ vue2/src/sample-app/main.ts
