+++ vue2/src/sample-app/router-async.d.ts
