+++ vue2/src/sample-app/router-async.ts
