+++ vue2/src/sample-app/router.d.ts
