+++ vue2/src/sample-app/shims-vue.d.ts
