+++ vue2/src/sample-app/views/About.vue
