+++ vue2/src/sample-app/views/Home.vue
